src/ai-orchestrator/requirements.txt
python-dotenv==1.0.1
azure-identity==1.17.1
openai==1.42.0
pydantic==2.8.2
streamlit==1.37.1
llama-index==0.10.68
llama-index-llms-openai==0.1.31
llama-index-llms-azure-openai==0.1.10
llama-index-program-openai==0.1.7
llama-index-agent-openai==0.2.9
llama-index-embeddings-openai==0.1.11
llama-index-agent-openai==0.2.9
llama-agents==0.0.14
aiokafka==0.11.0
kafka-python-ng==2.2.2